==> 4_BootstrappedMeans.Rmd <==
---
title: "Bootstrapped Moments"
author: "Mia Akbar"
date: "`r Sys.Date()`"
output: html_document
---


```{r setup}
library(tidyr)
library(dplyr)
library(ggplot2)
source("http://bit.ly/theme_pub")
theme_set(theme_pub()) 
setwd("./")
```

# Load data

Floral display data for all individuals

```{r}
Disp <- read.csv("./Data/Disp.csv")
```


Central moments of individual and aggregated population flowering schedules 

```{r moments}
Moments <- read.csv("./Data/Moments.csv")
AgMoments <- read.csv("./Data/AgMoments.csv")
```

Plant seed family and population information

```{r popinfo}
Pops <- read.csv("./Data/PopInfo.csv")
```

# Key Functions

Basic statistics:

First day that at least one flower is observed

```{r start_function}
startFlwr<-function(Days=NA, NF=NA){return(
  min(Days[NF>0],na.rm=T)
)}
```

Last day that at least one flower is observed

```{r end_function}
endFlwr<-function(Days=NA, NF=NA){return(
  max(Days[NF>0],na.rm=T)
)}
```

Duration of flowering

```{r duration_function}
FlwrDuration<-function(Days=NA, NF=NA){return(
  endFlwr(Days,NF)-startFlwr(Days,NF)
)}
```


Given a vector NF of flower numbers by day:

1. Vector of proportion of flowers

```{r Flwr}
pFlwr<-function(NF=NA){return(
  NF/sum(NF, na.rm=T)
  )}
```

2. Vector of days weighted by the proportion of flowers on each day

```{r WeightDay}
WeightDay<-function(Days=NA, NF=NA){return(
  Days*pFlwr(NF)
  )}
```

3. Weighted mean flowering day

```{r MeanDay}
MeanDay<-function(Days=NA, NF=NA){return(
  sum(WeightDay(Days,NF),na.rm=T)
)}
```

4. Weighted variance

```{r Var}
VarDay<-function(Days=NA, NF=NA){return(
  sum((Days-MeanDay(Days,NF))^2*pFlwr(NF),na.rm=T)
)}
```

5. Coefficient of variance

```{r CV}
CoefVarDay<-function(Days=NA, NF=NA){return(
  VarDay(Days,NF)/MeanDay(Days,NF)
)}
```

6. Coefficient of skewness

```{r Cskew}
CoefSkewDay<-function(Days=NA, NF=NA){return(
  sum((Days-MeanDay(Days,NF))^3*pFlwr(NF),na.rm=T)/
    sqrt(VarDay(Days,NF))^3
)}
```

7. Coefficient of excess kurtosis

```{r Ckurt}
CoefKurtDay<-function(Days=NA, NF=NA){return(
  sum((Days-MeanDay(Days,NF))^4*pFlwr(NF),na.rm=T)/
    sqrt(VarDay(Days,NF))^4 - 3
)}
```


# NEW Bootstrap models

The code generates 999 resampling events, and saves the output of the iterations for graphing, model fitting, and permutation testing below.

```{r}
Iters<-999 # Number of bootstrap iterations

# Raw Data
IndStats<-left_join(Moments,Pops,by="Ind") %>% select(!c(Ind,Fam,Pop))
PopDat<-Disp
PopLats<-unique(PopDat$Lat)

# Bootstrap output data frame
BootOut<-tibble(Iter=NA,Start=NA,Mean=NA,Var=NA,
                CV=NA,Skew=NA,Kurtosis=NA,End=NA,
                Duration=NA,Lat=NA)

Row<-1
for(i in 1:Iters){
  # Define a vector label for this iteration
  Iter<-rep(i,20)
  for(pop in 1:length(PopLats)){
    # Find the 20 individuals for 'pop'
    TempPop<-IndStats[IndStats$Lat==PopLats[pop],]    
    # Resample data for 20 individuals, with replacement
    # label with iteration (Iter)
    # and save to output dataframe (BootOut)
    BootOut[Row:(Row+19),]<-cbind(Iter,
                        TempPop[sample(c(1:20),20,replace=T),])
    Row<-Row+20
    TempPop<-NA
  }
  Iter<-NA
}

write.csv(BootOut,"./Data/BootOut.csv",row.names=F)
```

## Bootstrap mean and CI

This code calculates the bootstrap means and CIs for each population, and includes a Bonferoni correction for multiple tests (N=12) for 6 traits x 2 parameters (linear + quadratic).

```{r}
# Function for quantile calculations, with bonferroni correction
SumStats<-function(x=NA,bonfer=12){
  return(c(mean(x),quantile(x,
                  probs=c(0.005,0.025,0.975,0.99)/bonfer)))
}

# Calculate iteration averages
BootMeans<-BootOut %>%
  group_by(Lat,Iter)  %>%
  summarize(Start=mean(Start),Mean=mean(Mean),Var=mean(Var),CV=mean(CV),
            Skew=mean(Skew),Kurtosis=mean(Kurtosis),
            End=mean(End),Duration=mean(Duration)) %>%
  pivot_longer(cols=c(Start:Duration),
               names_to="Trait",
               values_to="IterMean")

SumStats<-BootMeans %>%
  group_by(Trait,Lat) %>%
  summarize(Mean=mean(IterMean),
            CI1=quantile(IterMean,probs=0.005),
            CI5=quantile(IterMean,probs=0.025),
            CI95=quantile(IterMean,probs=0.975),
            CI99=quantile(IterMean,probs=0.995))

write.csv(BootMeans,"./Data/BootMeans.csv",row.names=F)
write.csv(SumStats,"./Data/PopSummaryStats.csv",row.names=F)
```

## Bootstrap lm()

This code applies linear models to each bootstrap iteration

```{r}
BootMod<-tibble(Trait=NA,Iter=NA,Int=NA,Slope=NA,Quad=NA)

Traits<-unique(BootMeans$Trait)
Row<-1
for(trait in Traits){
  for(iter in 1:Iters){
    TempDat<-BootMeans[BootMeans$Trait==trait &   
                BootMeans$Iter==iter,]
    BootMod[Row,]<-t(c(trait,iter,
                     lm(IterMean~poly(Lat,2),data=TempDat)$coef))

    TempDat<-NA
    Row<-Row+1
    }
}
BootMod$Int<-as.numeric(BootMod$Int)
BootMod$Slope<-as.numeric(BootMod$Slope)
BootMod$Quad<-as.numeric(BootMod$Quad)

BootFits<-BootMod  %>%
  pivot_longer(cols=c(Int:Quad),
               names_to="Estimate",
               values_to="IterMean") %>%
  group_by(Trait,Estimate) %>%
  summarize(Mean=mean(IterMean),
            CI1=quantile(IterMean,probs=0.005),
            CI5=quantile(IterMean,probs=0.025),
            CI95=quantile(IterMean,probs=0.975),
            CI99=quantile(IterMean,probs=0.995))

write.csv(BootMod,"./Data/BootMod.csv")
write.csv(BootFits,"./Data/BootFits.csv")

```

## Bootstrap graphs

This code visualizes the 'best' linear model fit to bootstrap means as well as a subsample of bootstrap iterations. A random sample of 99 iterations are graphed to avoid unnecessary clutter.

```{r bootgraph_vis, fig.height=8,fig.width=6}
BootMeans<-read.csv("./Data/BootMeans.csv")
GraphDat<-BootMeans[grep("CV|End",BootMeans$Trait,invert=T),]
GraphDat<-GraphDat[GraphDat$Iter %in% c(1:99),]
MeanDat<-GraphDat %>%
  group_by(Trait,Lat)  %>%
  summarize(PopMean=mean(IterMean))


ggplot(aes(x=Lat,y=IterMean),data=GraphDat) +
  geom_point(colour=rgb(0,0,0,0.01),aes(group=Iter)) + 
  facet_wrap(vars(Trait),scales="free") + 
  geom_smooth(method="lm",formula = y ~ x + I(x^2),se=F,
              colour=rgb(0,0,0,0.01),aes(group=Iter)) +
  geom_point(aes(x=Lat,y=PopMean),data=MeanDat,
             colour="#000000") +
  geom_smooth(method="lm",formula = y ~ x + I(x^2),se=F,
              aes(x=Lat,y=PopMean),data=MeanDat,
             colour="#FF0000")
```
